WikiCommons

WikiCommons

wt_wikicommons(name, utf8 = TRUE, ...)

wt_wikicommons_parse(page, types = c("langlinks", "iwlinks", "externallinks",
  "common_names", "classification"), tidy = FALSE)

wt_wikicommons_search(query, limit = 10, offset = 0, utf8 = TRUE, ...)

Arguments

name

(character) Wiki name - as a page title

utf8

(logical) If `TRUE`, encodes most (but not all) non-ASCII characters as UTF-8 instead of replacing them with hexadecimal escape sequences. Default: `TRUE`

...

curl options, passed on to [httr::GET()]

page

([httr::response()]) Result of [wt_wiki_page()]

types

(character) List of properties to parse

tidy

(logical). tidy output to data.frame's if possible. Default: `FALSE`

query

(character) query terms

limit

(integer) number of results to return. Default: 10

offset

(integer) record to start at. Default: 0

Value

wt_wikicommons returns a list, with slots:

langlinks - language page links

externallinks - external links

common_names - a data.frame with name and language columns

classification - a data.frame with rank and name columns

wt_wikicommons_parse returns a list

wt_wikicommons_search returns a list with slots for continue and query, where query holds the results, with query$search slot with the search results

References

https://www.mediawiki.org/wiki/API:Search for help on search

Examples

# high level
wt_wikicommons(name = "Malus domestica")

#> $langlinks
#> # A tibble: 124 × 5
#>        lang
#> *     <chr>
#> 1       als
#> 2        am
#> 3        an
#> 4        ar
#> 5        as
#> 6       ast
#> 7        az
#> 8   bat-smg
#> 9        be
#> 10 be-x-old
#> # ... with 114 more rows, and 4 more variables: url <chr>, langname <chr>,
#> #   autonym <chr>, `*` <chr>
#> 
#> $externallinks
#> [1] "https://www.wikidata.org/wiki/Q158657"
#> 
#> $common_names
#> # A tibble: 151 × 2
#>              name language
#>             <chr>    <chr>
#> 1  paradise apple       en
#> 2           تفّاحة aeb-arab
#> 3         toffè7a aeb-latn
#> 4     Kulturapfel      als
#> 5              ፖም       am
#> 6          Mazana       an
#> 7        تفاح قزم       ar
#> 8            تفاح      arz
#> 9            আপেল       as
#> 10         Mazana      ast
#> # ... with 141 more rows
#> 
#> $classification
#> # A tibble: 16 × 2
#>            rank            name
#>           <chr>           <chr>
#> 1        Domain       Eukaryota
#> 2    • unranked  Archaeplastida
#> 3      • Regnum         Plantae
#> 4      • Cladus     angiosperms
#> 5      • Cladus        eudicots
#> 6      • Cladus   core eudicots
#> 7      • Cladus     superrosids
#> 8      • Cladus          rosids
#> 9      • Cladus      eurosids I
#> 10       • Ordo         Rosales
#> 11    • Familia        Rosaceae
#> 12 • Subfamilia    Spiraeoideae
#> 13     • Tribus          Pyreae
#> 14      • Genus           Malus
#> 15            • Malus domestica
#> 16                Borkh. (1803)
#> 

# low level
pg <- wt_wiki_page("https://commons.wikimedia.org/wiki/Malus_domestica")
wt_wikicommons_parse(pg)

#> $langlinks
#>   [1] "https://als.wikipedia.org/wiki/Kulturapfel"                                                               
#>   [2] "https://am.wikipedia.org/wiki/%E1%8D%96%E1%88%9D"                                                         
#>   [3] "https://an.wikipedia.org/wiki/Mazana"                                                                     
#>   [4] "https://ar.wikipedia.org/wiki/%D8%AA%D9%81%D8%A7%D8%AD_%D9%82%D8%B2%D9%85"                                
#>   [5] "https://as.wikipedia.org/wiki/%E0%A6%86%E0%A6%AA%E0%A7%87%E0%A6%B2"                                       
#>   [6] "https://ast.wikipedia.org/wiki/Mazana"                                                                    
#>   [7] "https://az.wikipedia.org/wiki/Alma_(meyv%C9%99)"                                                          
#>   [8] "https://bat-smg.wikipedia.org/wiki/Vuobel%C4%97s"                                                         
#>   [9] "https://be.wikipedia.org/wiki/%D0%AF%D0%B1%D0%BB%D1%8B%D0%BA"                                             
#>  [10] "https://be-x-old.wikipedia.org/wiki/%D0%AF%D0%B1%D0%BB%D1%8B%D0%BA"                                       
#>  [11] "https://bh.wikipedia.org/wiki/%E0%A4%B8%E0%A5%87%E0%A4%B5"                                                
#>  [12] "https://bo.wikipedia.org/wiki/%E0%BD%80%E0%BD%B4%E0%BC%8B%E0%BD%A4%E0%BD%B4%E0%BC%8D"                     
#>  [13] "https://bpy.wikipedia.org/wiki/%E0%A6%86%E0%A6%AA%E0%A7%87%E0%A6%B2"                                      
#>  [14] "https://br.wikipedia.org/wiki/Aval"                                                                       
#>  [15] "https://ca.wikipedia.org/wiki/Poma"                                                                       
#>  [16] "https://cdo.wikipedia.org/wiki/B%C3%ACng-gu%C5%8D"                                                        
#>  [17] "https://ceb.wikipedia.org/wiki/Malus_pumila"                                                              
#>  [18] "https://ckb.wikipedia.org/wiki/%D8%B3%DB%8E%D9%88"                                                        
#>  [19] "https://cs.wikipedia.org/wiki/Jablko"                                                                     
#>  [20] "https://csb.wikipedia.org/wiki/Dom%C3%B4c%C3%B4_jab%C5%82%C3%B3nka"                                       
#>  [21] "https://cv.wikipedia.org/wiki/%D0%9F%D0%B0%D0%BD%D1%83%D0%BB%D0%BC%D0%B8"                                 
#>  [22] "https://da.wikipedia.org/wiki/%C3%86ble"                                                                  
#>  [23] "https://de.wikipedia.org/wiki/Kulturapfel#Fr.C3.BCchte"                                                   
#>  [24] "https://dsb.wikipedia.org/wiki/Jab%C5%82uko"                                                              
#>  [25] "https://dv.wikipedia.org/wiki/%DE%87%DE%A7%DE%8A%DE%A6%DE%8D%DE%B0"                                       
#>  [26] "https://eml.wikipedia.org/wiki/P%C3%B9m"                                                                  
#>  [27] "https://eo.wikipedia.org/wiki/Pomo"                                                                       
#>  [28] "https://es.wikipedia.org/wiki/Manzana"                                                                    
#>  [29] "https://et.wikipedia.org/wiki/%C3%95un"                                                                   
#>  [30] "https://eu.wikipedia.org/wiki/Sagar"                                                                      
#>  [31] "https://fi.wikipedia.org/wiki/Tarhaomenapuu"                                                              
#>  [32] "https://fiu-vro.wikipedia.org/wiki/Uibu"                                                                  
#>  [33] "https://fj.wikipedia.org/wiki/Yapolo"                                                                     
#>  [34] "https://fr.wikipedia.org/wiki/Pomme"                                                                      
#>  [35] "https://frr.wikipedia.org/wiki/Aapelbuum"                                                                 
#>  [36] "https://fur.wikipedia.org/wiki/Milu%C3%A7%C3%A2r"                                                         
#>  [37] "https://ga.wikipedia.org/wiki/%C3%9All"                                                                   
#>  [38] "https://gan.wikipedia.org/wiki/%E8%98%8B%E6%9E%9C"                                                        
#>  [39] "https://gd.wikipedia.org/wiki/Ubhal"                                                                      
#>  [40] "https://gl.wikipedia.org/wiki/Maceira"                                                                    
#>  [41] "https://gu.wikipedia.org/wiki/%E0%AA%B8%E0%AA%AB%E0%AA%B0%E0%AA%9C%E0%AA%A8"                              
#>  [42] "https://gv.wikipedia.org/wiki/Ooyl"                                                                       
#>  [43] "https://hak.wikipedia.org/wiki/Ph%C3%ACn-k%C3%B3"                                                         
#>  [44] "https://he.wikipedia.org/wiki/%D7%AA%D7%A4%D7%95%D7%97"                                                   
#>  [45] "https://hi.wikipedia.org/wiki/%E0%A4%B8%E0%A5%87%E0%A4%AC"                                                
#>  [46] "https://hif.wikipedia.org/wiki/Aapul"                                                                     
#>  [47] "https://hsb.wikipedia.org/wiki/Jab%C5%82uko"                                                              
#>  [48] "https://ht.wikipedia.org/wiki/P%C3%B2m"                                                                   
#>  [49] "https://hu.wikipedia.org/wiki/Alma"                                                                       
#>  [50] "https://hy.wikipedia.org/wiki/%D4%BD%D5%B6%D5%B1%D5%B8%D6%80"                                             
#>  [51] "https://ia.wikipedia.org/wiki/Pomo"                                                                       
#>  [52] "https://it.wikipedia.org/wiki/Mela"                                                                       
#>  [53] "https://iu.wikipedia.org/wiki/%E1%91%AD%E1%92%BB%E1%92%A5%E1%93%87%E1%90%85%E1%94%AD%E1%96%85/kimminaujaq"
#>  [54] "https://jbo.wikipedia.org/wiki/plise"                                                                     
#>  [55] "https://jv.wikipedia.org/wiki/Apel"                                                                       
#>  [56] "https://kaa.wikipedia.org/wiki/Alma"                                                                      
#>  [57] "https://kk.wikipedia.org/wiki/%D0%90%D0%BB%D0%BC%D0%B0"                                                   
#>  [58] "https://kn.wikipedia.org/wiki/%E0%B2%B8%E0%B3%87%E0%B2%AC%E0%B3%81"                                       
#>  [59] "https://krc.wikipedia.org/wiki/%D0%90%D0%BB%D0%BC%D0%B0"                                                  
#>  [60] "https://ksh.wikipedia.org/wiki/Appel"                                                                     
#>  [61] "https://ku.wikipedia.org/wiki/S%C3%AAv"                                                                   
#>  [62] "https://kw.wikipedia.org/wiki/Aval"                                                                       
#>  [63] "https://la.wikipedia.org/wiki/Malum"                                                                      
#>  [64] "https://lb.wikipedia.org/wiki/Apel"                                                                       
#>  [65] "https://li.wikipedia.org/wiki/Appel"                                                                      
#>  [66] "https://lmo.wikipedia.org/wiki/Malus_communis"                                                            
#>  [67] "https://lt.wikipedia.org/wiki/Obuolys"                                                                    
#>  [68] "https://lv.wikipedia.org/wiki/%C4%80bols"                                                                 
#>  [69] "https://map-bms.wikipedia.org/wiki/Apel"                                                                  
#>  [70] "https://mg.wikipedia.org/wiki/Paoma"                                                                      
#>  [71] "https://mk.wikipedia.org/wiki/%D0%88%D0%B0%D0%B1%D0%BE%D0%BB%D0%BA%D0%BE"                                 
#>  [72] "https://mn.wikipedia.org/wiki/%D0%90%D0%BB%D0%B8%D0%BC"                                                   
#>  [73] "https://mr.wikipedia.org/wiki/%E0%A4%B8%E0%A4%AB%E0%A4%B0%E0%A4%9A%E0%A4%82%E0%A4%A6"                     
#>  [74] "https://ms.wikipedia.org/wiki/Epal"                                                                       
#>  [75] "https://nah.wikipedia.org/wiki/Manzana"                                                                   
#>  [76] "https://nap.wikipedia.org/wiki/Mir"                                                                       
#>  [77] "https://nds.wikipedia.org/wiki/Huusappel"                                                                 
#>  [78] "https://nds-nl.wikipedia.org/wiki/Appel"                                                                  
#>  [79] "https://ne.wikipedia.org/wiki/%E0%A4%B8%E0%A5%8D%E0%A4%AF%E0%A4%BE%E0%A4%89"                              
#>  [80] "https://new.wikipedia.org/wiki/%E0%A4%B8%E0%A5%8D%E0%A4%AF%E0%A4%BE%E0%A4%89"                             
#>  [81] "https://nl.wikipedia.org/wiki/Appel_(plant)"                                                              
#>  [82] "https://nn.wikipedia.org/wiki/Eple"                                                                       
#>  [83] "https://no.wikipedia.org/wiki/Epleslekten"                                                                
#>  [84] "https://nv.wikipedia.org/wiki/Bilas%C3%A1ana"                                                             
#>  [85] "https://oc.wikipedia.org/wiki/Poma"                                                                       
#>  [86] "https://os.wikipedia.org/wiki/%D0%A4%C3%A6%D1%82%D0%BA%D1%8A%D1%83%D1%8B"                                 
#>  [87] "https://pam.wikipedia.org/wiki/Mansanas"                                                                  
#>  [88] "https://pl.wikipedia.org/wiki/Jab%C5%82ko"                                                                
#>  [89] "https://pms.wikipedia.org/wiki/Malus_domestica"                                                           
#>  [90] "https://ps.wikipedia.org/wiki/%D9%85%DA%BC%D9%87"                                                         
#>  [91] "https://pt.wikipedia.org/wiki/Ma%C3%A7%C3%A3"                                                             
#>  [92] "https://qu.wikipedia.org/wiki/Mansana"                                                                    
#>  [93] "https://ro.wikipedia.org/wiki/M%C4%83r"                                                                   
#>  [94] "https://ru.wikipedia.org/wiki/%D0%AF%D0%B1%D0%BB%D0%BE%D0%BA%D0%BE"                                       
#>  [95] "https://rw.wikipedia.org/wiki/Pome"                                                                       
#>  [96] "https://sc.wikipedia.org/wiki/Mela"                                                                       
#>  [97] "https://scn.wikipedia.org/wiki/Pumu"                                                                      
#>  [98] "https://sco.wikipedia.org/wiki/Aiple"                                                                     
#>  [99] "https://sd.wikipedia.org/wiki/%D8%A7%D9%8E%D9%86%D8%A8%D9%90%DA%99%D9%90%D9%8A"                           
#> [100] "https://sh.wikipedia.org/wiki/Jabuka"                                                                     
#> [101] "https://si.wikipedia.org/wiki/%E0%B6%87%E0%B6%B4%E0%B6%BD%E0%B7%8A"                                       
#> [102] "https://simple.wikipedia.org/wiki/Apple"                                                                  
#> [103] "https://sk.wikipedia.org/wiki/Jablko"                                                                     
#> [104] "https://sl.wikipedia.org/wiki/Jabolko"                                                                    
#> [105] "https://sq.wikipedia.org/wiki/Molla"                                                                      
#> [106] "https://sr.wikipedia.org/wiki/%D0%88%D0%B0%D0%B1%D1%83%D0%BA%D0%B0"                                       
#> [107] "https://sv.wikipedia.org/wiki/%C3%84ppeltr%C3%A4d"                                                        
#> [108] "https://sw.wikipedia.org/wiki/Tofaa"                                                                      
#> [109] "https://te.wikipedia.org/wiki/%E0%B0%86%E0%B0%AA%E0%B0%BF%E0%B0%B2%E0%B1%8D"                              
#> [110] "https://tg.wikipedia.org/wiki/%D0%A1%D0%B5%D0%B1"                                                         
#> [111] "https://tl.wikipedia.org/wiki/Mansanas"                                                                   
#> [112] "https://tt.wikipedia.org/wiki/%D0%90%D0%BB%D0%BC%D0%B0"                                                   
#> [113] "https://udm.wikipedia.org/wiki/%D0%A3%D0%BB%D0%BC%D0%BE%D0%BF%D1%83"                                      
#> [114] "https://ug.wikipedia.org/wiki/%D8%A6%D8%A7%D9%84%D9%85%D8%A7"                                             
#> [115] "https://uk.wikipedia.org/wiki/%D0%AF%D0%B1%D0%BB%D1%83%D0%BA%D0%BE"                                       
#> [116] "https://uz.wikipedia.org/wiki/Olma"                                                                       
#> [117] "https://ve.wikipedia.org/wiki/Apula"                                                                      
#> [118] "https://vec.wikipedia.org/wiki/Pomaro"                                                                    
#> [119] "https://vi.wikipedia.org/wiki/Malus_domestica"                                                            
#> [120] "https://wa.wikipedia.org/wiki/Pem%C3%AE"                                                                  
#> [121] "https://war.wikipedia.org/wiki/Mansanas"                                                                  
#> [122] "https://yi.wikipedia.org/wiki/%D7%A2%D7%A4%D7%9C"                                                         
#> [123] "https://zh-min-nan.wikipedia.org/wiki/Ph%C5%8Dng-k%C3%B3-chhi%C5%AB"                                      
#> [124] "https://zh-yue.wikipedia.org/wiki/%E8%98%8B%E6%9E%9C"                                                     
#> 
#> $iwlinks
#>  [1] "https://en.wikipedia.org/wiki/Wikispecies:"                                                                               
#>  [2] "https://en.wikipedia.org/wiki/Common_name"                                                                                
#>  [3] "https://species.wikimedia.org/wiki/Malus_domestica"                                                                       
#>  [4] "https://www.wikidata.org/wiki/Q158657"                                                                                    
#>  [5] "https://www.wikidata.org/wiki/P935"                                                                                       
#>  [6] "https://als.wikipedia.org/wiki/Kulturapfel"                                                                               
#>  [7] "https://am.wikipedia.org/wiki/%E1%8D%96%E1%88%9D"                                                                         
#>  [8] "https://ar.wikipedia.org/wiki/%D8%AA%D9%81%D8%A7%D8%AD_%D9%82%D8%B2%D9%85"                                                
#>  [9] "https://ast.wikipedia.org/wiki/Malus_domestica"                                                                           
#> [10] "https://az.wikipedia.org/wiki/M%C9%99d%C9%99ni_alma_a%C4%9Fac%C4%B1"                                                      
#> [11] "https://be.wikipedia.org/wiki/Malus_pumila"                                                                               
#> [12] "https://bo.wikipedia.org/wiki/%E0%BD%80%E0%BD%B4%E0%BC%8B%E0%BD%A4%E0%BD%B4%E0%BC%8D"                                     
#> [13] "https://bpy.wikipedia.org/wiki/%E0%A6%86%E0%A6%AA%E0%A7%87%E0%A6%B2"                                                      
#> [14] "https://ca.wikipedia.org/wiki/Pomera_comuna"                                                                              
#> [15] "https://cdo.wikipedia.org/wiki/B%C3%ACng-gu%C5%8D"                                                                        
#> [16] "https://cs.wikipedia.org/wiki/Jablo%C5%88_dom%C3%A1c%C3%AD"                                                               
#> [17] "https://csb.wikipedia.org/wiki/Dom%C3%B4c%C3%B4_jab%C5%82%C3%B3nka"                                                       
#> [18] "https://da.wikipedia.org/wiki/Almindelig_%C3%A6ble"                                                                       
#> [19] "https://de.wikipedia.org/wiki/Kulturapfel"                                                                                
#> [20] "https://dsb.wikipedia.org/wiki/Zagrodna_jab%C5%82u%C5%A1cyna"                                                             
#> [21] "https://eml.wikipedia.org/wiki/P%C3%B9m"                                                                                  
#> [22] "https://eo.wikipedia.org/wiki/Kulturpomo"                                                                                 
#> [23] "https://es.wikipedia.org/wiki/Malus_domestica"                                                                            
#> [24] "https://et.wikipedia.org/wiki/%C3%95unapuu"                                                                               
#> [25] "https://eu.wikipedia.org/wiki/Sagarrondo"                                                                                 
#> [26] "https://fi.wikipedia.org/wiki/Tarhaomenapuu"                                                                              
#> [27] "https://fr.wikipedia.org/wiki/Pommier_domestique"                                                                         
#> [28] "https://frr.wikipedia.org/wiki/Aapelbuum"                                                                                 
#> [29] "https://gl.wikipedia.org/wiki/Maceira"                                                                                    
#> [30] "https://gv.wikipedia.org/wiki/Ooyl"                                                                                       
#> [31] "https://hsb.wikipedia.org/wiki/Zahrodna_jab%C5%82u%C4%8Dina"                                                              
#> [32] "https://hu.wikipedia.org/wiki/Nemes_alma"                                                                                 
#> [33] "https://hy.wikipedia.org/wiki/%D4%BD%D5%B6%D5%B1%D5%B8%D6%80"                                                             
#> [34] "https://ia.wikipedia.org/wiki/Malus_pumila"                                                                               
#> [35] "https://it.wikipedia.org/wiki/Malus_domestica"                                                                            
#> [36] "https://jv.wikipedia.org/wiki/Apel"                                                                                       
#> [37] "https://ku.wikipedia.org/wiki/S%C3%AAv"                                                                                   
#> [38] "https://kw.wikipedia.org/wiki/Aval"                                                                                       
#> [39] "https://la.wikipedia.org/wiki/Malus_pumila"                                                                               
#> [40] "https://li.wikipedia.org/wiki/Appel"                                                                                      
#> [41] "https://lmo.wikipedia.org/wiki/Malus_communis"                                                                            
#> [42] "https://lv.wikipedia.org/wiki/%C4%80bele"                                                                                 
#> [43] "https://mk.wikipedia.org/wiki/%D0%88%D0%B0%D0%B1%D0%BE%D0%BB%D0%BA%D0%BE"                                                 
#> [44] "https://mn.wikipedia.org/wiki/%D0%90%D0%BB%D0%B8%D0%BC"                                                                   
#> [45] "https://ms.wikipedia.org/wiki/Epal"                                                                                       
#> [46] "https://nah.wikipedia.org/wiki/Manzanacuahuitl"                                                                           
#> [47] "https://nap.wikipedia.org/wiki/Mir"                                                                                       
#> [48] "https://nds.wikipedia.org/wiki/Huusappel"                                                                                 
#> [49] "https://ne.wikipedia.org/wiki/%E0%A4%B8%E0%A5%8D%E0%A4%AF%E0%A4%BE%E0%A4%89"                                              
#> [50] "https://nl.wikipedia.org/wiki/Appel_(plant)"                                                                              
#> [51] "https://nn.wikipedia.org/wiki/Eple"                                                                                       
#> [52] "https://oc.wikipedia.org/wiki/Pomi%C3%A8r"                                                                                
#> [53] "https://os.wikipedia.org/wiki/%D0%A5%C3%A6%D0%B4%D0%B7%D0%B0%D1%80%D0%BE%D0%BD_%D1%84%C3%A6%D1%82%D0%BA%D1%8A%D1%83%D1%8B"
#> [54] "https://pam.wikipedia.org/wiki/Mansanas"                                                                                  
#> [55] "https://pl.wikipedia.org/wiki/Jab%C5%82o%C5%84_domowa"                                                                    
#> [56] "https://ps.wikipedia.org/wiki/%D9%85%DA%BC%D9%87"                                                                         
#> [57] "https://pt.wikipedia.org/wiki/Malus_domestica"                                                                            
#> [58] "https://qu.wikipedia.org/wiki/Mansana"                                                                                    
#> [59] "https://ro.wikipedia.org/wiki/M%C4%83r"                                                                                   
#> [60] "https://ru.wikipedia.org/wiki/%D0%AF%D0%B1%D0%BB%D0%BE%D0%BD%D1%8F_%D0%B4%D0%BE%D0%BC%D0%B0%D1%88%D0%BD%D1%8F%D1%8F"      
#> [61] "https://sc.wikipedia.org/wiki/Mela"                                                                                       
#> [62] "https://sh.wikipedia.org/wiki/Jabuka"                                                                                     
#> [63] "https://si.wikipedia.org/wiki/%E0%B6%87%E0%B6%B4%E0%B6%BD%E0%B7%8A"                                                       
#> [64] "https://simple.wikipedia.org/wiki/Apple"                                                                                  
#> [65] "https://sk.wikipedia.org/wiki/Jablo%C5%88_dom%C3%A1ca"                                                                    
#> [66] "https://sl.wikipedia.org/wiki/Jablana"                                                                                    
#> [67] "https://sr.wikipedia.org/wiki/%D0%94%D0%BE%D0%BC%D0%B0%D1%9B%D0%B0_%D1%98%D0%B0%D0%B1%D1%83%D0%BA%D0%B0"                  
#> [68] "https://sv.wikipedia.org/wiki/%C3%84ppeltr%C3%A4d"                                                                        
#> [69] "https://te.wikipedia.org/wiki/%E0%B0%86%E0%B0%AA%E0%B0%BF%E0%B0%B2%E0%B1%8D"                                              
#> [70] "https://uk.wikipedia.org/wiki/%D0%AF%D0%B1%D0%BB%D1%83%D0%BD%D1%8F_%D0%B4%D0%BE%D0%BC%D0%B0%D1%88%D0%BD%D1%8F"            
#> [71] "https://uz.wikipedia.org/wiki/Olma"                                                                                       
#> [72] "https://vec.wikipedia.org/wiki/Pomaro"                                                                                    
#> [73] "https://vi.wikipedia.org/wiki/Malus_domestica"                                                                            
#> [74] "https://war.wikipedia.org/wiki/Malus_pumila"                                                                              
#> 
#> $externallinks
#> $externallinks[[1]]
#> [1] "https://www.wikidata.org/wiki/Q158657"
#> 
#> 
#> $common_names
#> $common_names[[1]]
#> $common_names[[1]]$name
#> [1] "paradise apple"
#> 
#> $common_names[[1]]$language
#> [1] "en"
#> 
#> 
#> $common_names[[2]]
#> $common_names[[2]]$name
#> [1] "تفّاحة"
#> 
#> $common_names[[2]]$language
#> [1] "aeb-arab"
#> 
#> 
#> $common_names[[3]]
#> $common_names[[3]]$name
#> [1] "toffè7a"
#> 
#> $common_names[[3]]$language
#> [1] "aeb-latn"
#> 
#> 
#> $common_names[[4]]
#> $common_names[[4]]$name
#> [1] "Kulturapfel"
#> 
#> $common_names[[4]]$language
#> [1] "als"
#> 
#> 
#> $common_names[[5]]
#> $common_names[[5]]$name
#> [1] "ፖም"
#> 
#> $common_names[[5]]$language
#> [1] "am"
#> 
#> 
#> $common_names[[6]]
#> $common_names[[6]]$name
#> [1] "Mazana"
#> 
#> $common_names[[6]]$language
#> [1] "an"
#> 
#> 
#> $common_names[[7]]
#> $common_names[[7]]$name
#> [1] "تفاح قزم"
#> 
#> $common_names[[7]]$language
#> [1] "ar"
#> 
#> 
#> $common_names[[8]]
#> $common_names[[8]]$name
#> [1] "تفاح"
#> 
#> $common_names[[8]]$language
#> [1] "arz"
#> 
#> 
#> $common_names[[9]]
#> $common_names[[9]]$name
#> [1] "আপেল"
#> 
#> $common_names[[9]]$language
#> [1] "as"
#> 
#> 
#> $common_names[[10]]
#> $common_names[[10]]$name
#> [1] "Mazana"
#> 
#> $common_names[[10]]$language
#> [1] "ast"
#> 
#> 
#> $common_names[[11]]
#> $common_names[[11]]$name
#> [1] "Mədəni alma ağacı"
#> 
#> $common_names[[11]]$language
#> [1] "az"
#> 
#> 
#> $common_names[[12]]
#> $common_names[[12]]$name
#> [1] "Epfe"
#> 
#> $common_names[[12]]$language
#> [1] "bar"
#> 
#> 
#> $common_names[[13]]
#> $common_names[[13]]$name
#> [1] "Vuobelis"
#> 
#> $common_names[[13]]$language
#> [1] "bat-smg"
#> 
#> 
#> $common_names[[14]]
#> $common_names[[14]]$name
#> [1] "Яблык"
#> 
#> $common_names[[14]]$language
#> [1] "be"
#> 
#> 
#> $common_names[[15]]
#> $common_names[[15]]$name
#> [1] "Яблык"
#> 
#> $common_names[[15]]$language
#> [1] "be-x-old"
#> 
#> 
#> $common_names[[16]]
#> $common_names[[16]]$name
#> [1] "Ябълка"
#> 
#> $common_names[[16]]$language
#> [1] "bg"
#> 
#> 
#> $common_names[[17]]
#> $common_names[[17]]$name
#> [1] "सेव"
#> 
#> $common_names[[17]]$language
#> [1] "bh"
#> 
#> 
#> $common_names[[18]]
#> $common_names[[18]]$name
#> [1] "মালুস পুমিলা, আপেল"
#> 
#> $common_names[[18]]$language
#> [1] "bn"
#> 
#> 
#> $common_names[[19]]
#> $common_names[[19]]$name
#> [1] "ཀུ་ཤུ།"
#> 
#> $common_names[[19]]$language
#> [1] "bo"
#> 
#> 
#> $common_names[[20]]
#> $common_names[[20]]$name
#> [1] "আপেল"
#> 
#> $common_names[[20]]$language
#> [1] "bpy"
#> 
#> 
#> $common_names[[21]]
#> $common_names[[21]]$name
#> [1] "Aval"
#> 
#> $common_names[[21]]$language
#> [1] "br"
#> 
#> 
#> $common_names[[22]]
#> $common_names[[22]]$name
#> [1] "Jabuka"
#> 
#> $common_names[[22]]$language
#> [1] "bs"
#> 
#> 
#> $common_names[[23]]
#> $common_names[[23]]$name
#> [1] "Pomera comuna, Poma"
#> 
#> $common_names[[23]]$language
#> [1] "ca"
#> 
#> 
#> $common_names[[24]]
#> $common_names[[24]]$name
#> [1] "Bìng-guō"
#> 
#> $common_names[[24]]$language
#> [1] "cdo"
#> 
#> 
#> $common_names[[25]]
#> $common_names[[25]]$name
#> [1] "سێو"
#> 
#> $common_names[[25]]$language
#> [1] "ckb"
#> 
#> 
#> $common_names[[26]]
#> $common_names[[26]]$name
#> [1] "jabloň nízká, jabloň domácí"
#> 
#> $common_names[[26]]$language
#> [1] "cs"
#> 
#> 
#> $common_names[[27]]
#> $common_names[[27]]$name
#> [1] "Domôcô jabłónka"
#> 
#> $common_names[[27]]$language
#> [1] "csb"
#> 
#> 
#> $common_names[[28]]
#> $common_names[[28]]$name
#> [1] "Панулми"
#> 
#> $common_names[[28]]$language
#> [1] "cv"
#> 
#> 
#> $common_names[[29]]
#> $common_names[[29]]$name
#> [1] "Afal"
#> 
#> $common_names[[29]]$language
#> [1] "cy"
#> 
#> 
#> $common_names[[30]]
#> $common_names[[30]]$name
#> [1] "Almindelig æble, Æble"
#> 
#> $common_names[[30]]$language
#> [1] "da"
#> 
#> 
#> $common_names[[31]]
#> $common_names[[31]]$name
#> [1] "Kulturapfel"
#> 
#> $common_names[[31]]$language
#> [1] "de"
#> 
#> 
#> $common_names[[32]]
#> $common_names[[32]]$name
#> [1] "Zagrodna jabłušcyna, Jabłuko"
#> 
#> $common_names[[32]]$language
#> [1] "dsb"
#> 
#> 
#> $common_names[[33]]
#> $common_names[[33]]$name
#> [1] "އާފަލް"
#> 
#> $common_names[[33]]$language
#> [1] "dv"
#> 
#> 
#> $common_names[[34]]
#> $common_names[[34]]$name
#> [1] "μηλιά, Μήλο"
#> 
#> $common_names[[34]]$language
#> [1] "el"
#> 
#> 
#> $common_names[[35]]
#> $common_names[[35]]$name
#> [1] "Pùm, Pàm"
#> 
#> $common_names[[35]]$language
#> [1] "eml"
#> 
#> 
#> $common_names[[36]]
#> $common_names[[36]]$name
#> [1] "Kulturpomo"
#> 
#> $common_names[[36]]$language
#> [1] "eo"
#> 
#> 
#> $common_names[[37]]
#> $common_names[[37]]$name
#> [1] "manzano, Manzana"
#> 
#> $common_names[[37]]$language
#> [1] "es"
#> 
#> 
#> $common_names[[38]]
#> $common_names[[38]]$name
#> [1] "Madal õunapuu, Õun"
#> 
#> $common_names[[38]]$language
#> [1] "et"
#> 
#> 
#> $common_names[[39]]
#> $common_names[[39]]$name
#> [1] "Sagarrondo"
#> 
#> $common_names[[39]]$language
#> [1] "eu"
#> 
#> 
#> $common_names[[40]]
#> $common_names[[40]]$name
#> [1] "سیب"
#> 
#> $common_names[[40]]$language
#> [1] "fa"
#> 
#> 
#> $common_names[[41]]
#> $common_names[[41]]$name
#> [1] "Tarhaomenapuu"
#> 
#> $common_names[[41]]$language
#> [1] "fi"
#> 
#> 
#> $common_names[[42]]
#> $common_names[[42]]$name
#> [1] "Yapolo"
#> 
#> $common_names[[42]]$language
#> [1] "fj"
#> 
#> 
#> $common_names[[43]]
#> $common_names[[43]]$name
#> [1] "Pommier domestique, Pomme"
#> 
#> $common_names[[43]]$language
#> [1] "fr"
#> 
#> 
#> $common_names[[44]]
#> $common_names[[44]]$name
#> [1] "Aapelbuum"
#> 
#> $common_names[[44]]$language
#> [1] "frr"
#> 
#> 
#> $common_names[[45]]
#> $common_names[[45]]$name
#> [1] "Miluçâr"
#> 
#> $common_names[[45]]$language
#> [1] "fur"
#> 
#> 
#> $common_names[[46]]
#> $common_names[[46]]$name
#> [1] "Úll"
#> 
#> $common_names[[46]]$language
#> [1] "ga"
#> 
#> 
#> $common_names[[47]]
#> $common_names[[47]]$name
#> [1] "蘋果"
#> 
#> $common_names[[47]]$language
#> [1] "gan"
#> 
#> 
#> $common_names[[48]]
#> $common_names[[48]]$name
#> [1] "Ubhal"
#> 
#> $common_names[[48]]$language
#> [1] "gd"
#> 
#> 
#> $common_names[[49]]
#> $common_names[[49]]$name
#> [1] "Maceira"
#> 
#> $common_names[[49]]$language
#> [1] "gl"
#> 
#> 
#> $common_names[[50]]
#> $common_names[[50]]$name
#> [1] "Kulturapfel"
#> 
#> $common_names[[50]]$language
#> [1] "gsw"
#> 
#> 
#> $common_names[[51]]
#> $common_names[[51]]$name
#> [1] "સફરજન"
#> 
#> $common_names[[51]]$language
#> [1] "gu"
#> 
#> 
#> $common_names[[52]]
#> $common_names[[52]]$name
#> [1] "Ooyl"
#> 
#> $common_names[[52]]$language
#> [1] "gv"
#> 
#> 
#> $common_names[[53]]
#> $common_names[[53]]$name
#> [1] "Phìn-kó"
#> 
#> $common_names[[53]]$language
#> [1] "hak"
#> 
#> 
#> $common_names[[54]]
#> $common_names[[54]]$name
#> [1] "תפוח"
#> 
#> $common_names[[54]]$language
#> [1] "he"
#> 
#> 
#> $common_names[[55]]
#> $common_names[[55]]$name
#> [1] "सेब"
#> 
#> $common_names[[55]]$language
#> [1] "hi"
#> 
#> 
#> $common_names[[56]]
#> $common_names[[56]]$name
#> [1] "Aapul"
#> 
#> $common_names[[56]]$language
#> [1] "hif"
#> 
#> 
#> $common_names[[57]]
#> $common_names[[57]]$name
#> [1] "Jabuka"
#> 
#> $common_names[[57]]$language
#> [1] "hr"
#> 
#> 
#> $common_names[[58]]
#> $common_names[[58]]$name
#> [1] "Zahrodna jabłučina, Jabłuko"
#> 
#> $common_names[[58]]$language
#> [1] "hsb"
#> 
#> 
#> $common_names[[59]]
#> $common_names[[59]]$name
#> [1] "Pòm"
#> 
#> $common_names[[59]]$language
#> [1] "ht"
#> 
#> 
#> $common_names[[60]]
#> $common_names[[60]]$name
#> [1] "Nemes alma"
#> 
#> $common_names[[60]]$language
#> [1] "hu"
#> 
#> 
#> $common_names[[61]]
#> $common_names[[61]]$name
#> [1] "Խնձոր"
#> 
#> $common_names[[61]]$language
#> [1] "hy"
#> 
#> 
#> $common_names[[62]]
#> $common_names[[62]]$name
#> [1] "Pomo"
#> 
#> $common_names[[62]]$language
#> [1] "ia"
#> 
#> 
#> $common_names[[63]]
#> $common_names[[63]]$name
#> [1] "Apel"
#> 
#> $common_names[[63]]$language
#> [1] "id"
#> 
#> 
#> $common_names[[64]]
#> $common_names[[64]]$name
#> [1] "Epli"
#> 
#> $common_names[[64]]$language
#> [1] "is"
#> 
#> 
#> $common_names[[65]]
#> $common_names[[65]]$name
#> [1] "melo, Mela"
#> 
#> $common_names[[65]]$language
#> [1] "it"
#> 
#> 
#> $common_names[[66]]
#> $common_names[[66]]$name
#> [1] "ᑭᒻᒥᓇᐅᔭᖅ/kimminaujaq"
#> 
#> $common_names[[66]]$language
#> [1] "iu"
#> 
#> 
#> $common_names[[67]]
#> $common_names[[67]]$name
#> [1] "リンゴ"
#> 
#> $common_names[[67]]$language
#> [1] "ja"
#> 
#> 
#> $common_names[[68]]
#> $common_names[[68]]$name
#> [1] "plise"
#> 
#> $common_names[[68]]$language
#> [1] "jbo"
#> 
#> 
#> $common_names[[69]]
#> $common_names[[69]]$name
#> [1] "Apel"
#> 
#> $common_names[[69]]$language
#> [1] "jv"
#> 
#> 
#> $common_names[[70]]
#> $common_names[[70]]$name
#> [1] "Alma"
#> 
#> $common_names[[70]]$language
#> [1] "kaa"
#> 
#> 
#> $common_names[[71]]
#> $common_names[[71]]$name
#> [1] "Алма"
#> 
#> $common_names[[71]]$language
#> [1] "kk"
#> 
#> 
#> $common_names[[72]]
#> $common_names[[72]]$name
#> [1] "ಸೇಬು"
#> 
#> $common_names[[72]]$language
#> [1] "kn"
#> 
#> 
#> $common_names[[73]]
#> $common_names[[73]]$name
#> [1] "사과"
#> 
#> $common_names[[73]]$language
#> [1] "ko"
#> 
#> 
#> $common_names[[74]]
#> $common_names[[74]]$name
#> [1] "Алма"
#> 
#> $common_names[[74]]$language
#> [1] "krc"
#> 
#> 
#> $common_names[[75]]
#> $common_names[[75]]$name
#> [1] "Appel"
#> 
#> $common_names[[75]]$language
#> [1] "ksh"
#> 
#> 
#> $common_names[[76]]
#> $common_names[[76]]$name
#> [1] "Sêv"
#> 
#> $common_names[[76]]$language
#> [1] "ku"
#> 
#> 
#> $common_names[[77]]
#> $common_names[[77]]$name
#> [1] "Aval"
#> 
#> $common_names[[77]]$language
#> [1] "kw"
#> 
#> 
#> $common_names[[78]]
#> $common_names[[78]]$name
#> [1] "Malum"
#> 
#> $common_names[[78]]$language
#> [1] "la"
#> 
#> 
#> $common_names[[79]]
#> $common_names[[79]]$name
#> [1] "Apel"
#> 
#> $common_names[[79]]$language
#> [1] "lb"
#> 
#> 
#> $common_names[[80]]
#> $common_names[[80]]$name
#> [1] "Appel"
#> 
#> $common_names[[80]]$language
#> [1] "li"
#> 
#> 
#> $common_names[[81]]
#> $common_names[[81]]$name
#> [1] "Malus communis"
#> 
#> $common_names[[81]]$language
#> [1] "lmo"
#> 
#> 
#> $common_names[[82]]
#> $common_names[[82]]$name
#> [1] "Obuolys"
#> 
#> $common_names[[82]]$language
#> [1] "lt"
#> 
#> 
#> $common_names[[83]]
#> $common_names[[83]]$name
#> [1] "Ābele, Ābols"
#> 
#> $common_names[[83]]$language
#> [1] "lv"
#> 
#> 
#> $common_names[[84]]
#> $common_names[[84]]$name
#> [1] "Apel"
#> 
#> $common_names[[84]]$language
#> [1] "map-bms"
#> 
#> 
#> $common_names[[85]]
#> $common_names[[85]]$name
#> [1] "Paoma"
#> 
#> $common_names[[85]]$language
#> [1] "mg"
#> 
#> 
#> $common_names[[86]]
#> $common_names[[86]]$name
#> [1] "Јаболко"
#> 
#> $common_names[[86]]$language
#> [1] "mk"
#> 
#> 
#> $common_names[[87]]
#> $common_names[[87]]$name
#> [1] "ആപ്പി\u0d7e"
#> 
#> $common_names[[87]]$language
#> [1] "ml"
#> 
#> 
#> $common_names[[88]]
#> $common_names[[88]]$name
#> [1] "Алим"
#> 
#> $common_names[[88]]$language
#> [1] "mn"
#> 
#> 
#> $common_names[[89]]
#> $common_names[[89]]$name
#> [1] "सफरचंद"
#> 
#> $common_names[[89]]$language
#> [1] "mr"
#> 
#> 
#> $common_names[[90]]
#> $common_names[[90]]$name
#> [1] "Epal"
#> 
#> $common_names[[90]]$language
#> [1] "ms"
#> 
#> 
#> $common_names[[91]]
#> $common_names[[91]]$name
#> [1] "Manzanacuahuitl"
#> 
#> $common_names[[91]]$language
#> [1] "nah"
#> 
#> 
#> $common_names[[92]]
#> $common_names[[92]]$name
#> [1] "Mir"
#> 
#> $common_names[[92]]$language
#> [1] "nap"
#> 
#> 
#> $common_names[[93]]
#> $common_names[[93]]$name
#> [1] "Huusappel"
#> 
#> $common_names[[93]]$language
#> [1] "nds"
#> 
#> 
#> $common_names[[94]]
#> $common_names[[94]]$name
#> [1] "Appel"
#> 
#> $common_names[[94]]$language
#> [1] "nds-nl"
#> 
#> 
#> $common_names[[95]]
#> $common_names[[95]]$name
#> [1] "स्याउ"
#> 
#> $common_names[[95]]$language
#> [1] "ne"
#> 
#> 
#> $common_names[[96]]
#> $common_names[[96]]$name
#> [1] "स्याउ"
#> 
#> $common_names[[96]]$language
#> [1] "new"
#> 
#> 
#> $common_names[[97]]
#> $common_names[[97]]$name
#> [1] "Appel"
#> 
#> $common_names[[97]]$language
#> [1] "nl"
#> 
#> 
#> $common_names[[98]]
#> $common_names[[98]]$name
#> [1] "Eple"
#> 
#> $common_names[[98]]$language
#> [1] "nn"
#> 
#> 
#> $common_names[[99]]
#> $common_names[[99]]$name
#> [1] "Epleslekten"
#> 
#> $common_names[[99]]$language
#> [1] "no"
#> 
#> 
#> $common_names[[100]]
#> $common_names[[100]]$name
#> [1] "Bilasáana"
#> 
#> $common_names[[100]]$language
#> [1] "nv"
#> 
#> 
#> $common_names[[101]]
#> $common_names[[101]]$name
#> [1] "Pomièr, Poma"
#> 
#> $common_names[[101]]$language
#> [1] "oc"
#> 
#> 
#> $common_names[[102]]
#> $common_names[[102]]$name
#> [1] "Хæдзарон фæткъуы, Фæткъуы"
#> 
#> $common_names[[102]]$language
#> [1] "os"
#> 
#> 
#> $common_names[[103]]
#> $common_names[[103]]$name
#> [1] "Mansanas"
#> 
#> $common_names[[103]]$language
#> [1] "pam"
#> 
#> 
#> $common_names[[104]]
#> $common_names[[104]]$name
#> [1] "Jabłoń domowa, Jabłko"
#> 
#> $common_names[[104]]$language
#> [1] "pl"
#> 
#> 
#> $common_names[[105]]
#> $common_names[[105]]$name
#> [1] "سیب"
#> 
#> $common_names[[105]]$language
#> [1] "pnb"
#> 
#> 
#> $common_names[[106]]
#> $common_names[[106]]$name
#> [1] "مڼه"
#> 
#> $common_names[[106]]$language
#> [1] "ps"
#> 
#> 
#> $common_names[[107]]
#> $common_names[[107]]$name
#> [1] "Maçã"
#> 
#> $common_names[[107]]$language
#> [1] "pt"
#> 
#> 
#> $common_names[[108]]
#> $common_names[[108]]$name
#> [1] "Mansana"
#> 
#> $common_names[[108]]$language
#> [1] "qu"
#> 
#> 
#> $common_names[[109]]
#> $common_names[[109]]$name
#> [1] "Măr"
#> 
#> $common_names[[109]]$language
#> [1] "ro"
#> 
#> 
#> $common_names[[110]]
#> $common_names[[110]]$name
#> [1] "Яблоня домашняя, Яблоко"
#> 
#> $common_names[[110]]$language
#> [1] "ru"
#> 
#> 
#> $common_names[[111]]
#> $common_names[[111]]$name
#> [1] "Pome"
#> 
#> $common_names[[111]]$language
#> [1] "rw"
#> 
#> 
#> $common_names[[112]]
#> $common_names[[112]]$name
#> [1] "Mela"
#> 
#> $common_names[[112]]$language
#> [1] "sc"
#> 
#> 
#> $common_names[[113]]
#> $common_names[[113]]$name
#> [1] "pumu"
#> 
#> $common_names[[113]]$language
#> [1] "scn"
#> 
#> 
#> $common_names[[114]]
#> $common_names[[114]]$name
#> [1] "aiple tree"
#> 
#> $common_names[[114]]$language
#> [1] "sco"
#> 
#> 
#> $common_names[[115]]
#> $common_names[[115]]$name
#> [1] "اَنبِڙِي"
#> 
#> $common_names[[115]]$language
#> [1] "sd"
#> 
#> 
#> $common_names[[116]]
#> $common_names[[116]]$name
#> [1] "Vuobelės"
#> 
#> $common_names[[116]]$language
#> [1] "sgs"
#> 
#> 
#> $common_names[[117]]
#> $common_names[[117]]$name
#> [1] "Jabuka"
#> 
#> $common_names[[117]]$language
#> [1] "sh"
#> 
#> 
#> $common_names[[118]]
#> $common_names[[118]]$name
#> [1] "ඇපල්"
#> 
#> $common_names[[118]]$language
#> [1] "si"
#> 
#> 
#> $common_names[[119]]
#> $common_names[[119]]$name
#> [1] "Apple"
#> 
#> $common_names[[119]]$language
#> [1] "simple"
#> 
#> 
#> $common_names[[120]]
#> $common_names[[120]]$name
#> [1] "Jabloň domáca, Jablko"
#> 
#> $common_names[[120]]$language
#> [1] "sk"
#> 
#> 
#> $common_names[[121]]
#> $common_names[[121]]$name
#> [1] "Jablana"
#> 
#> $common_names[[121]]$language
#> [1] "sl"
#> 
#> 
#> $common_names[[122]]
#> $common_names[[122]]$name
#> [1] "Tufaax"
#> 
#> $common_names[[122]]$language
#> [1] "so"
#> 
#> 
#> $common_names[[123]]
#> $common_names[[123]]$name
#> [1] "Molla"
#> 
#> $common_names[[123]]$language
#> [1] "sq"
#> 
#> 
#> $common_names[[124]]
#> $common_names[[124]]$name
#> [1] "Домаћа јабука, Јабука"
#> 
#> $common_names[[124]]$language
#> [1] "sr"
#> 
#> 
#> $common_names[[125]]
#> $common_names[[125]]$name
#> [1] "Домаћа јабука"
#> 
#> $common_names[[125]]$language
#> [1] "sr-ec"
#> 
#> 
#> $common_names[[126]]
#> $common_names[[126]]$name
#> [1] "Domaća jabuka"
#> 
#> $common_names[[126]]$language
#> [1] "sr-el"
#> 
#> 
#> $common_names[[127]]
#> $common_names[[127]]$name
#> [1] "Äppelträd, Äpple"
#> 
#> $common_names[[127]]$language
#> [1] "sv"
#> 
#> 
#> $common_names[[128]]
#> $common_names[[128]]$name
#> [1] "Tofaa"
#> 
#> $common_names[[128]]$language
#> [1] "sw"
#> 
#> 
#> $common_names[[129]]
#> $common_names[[129]]$name
#> [1] "ஆப்பிள்"
#> 
#> $common_names[[129]]$language
#> [1] "ta"
#> 
#> 
#> $common_names[[130]]
#> $common_names[[130]]$name
#> [1] "ఆపిల్"
#> 
#> $common_names[[130]]$language
#> [1] "te"
#> 
#> 
#> $common_names[[131]]
#> $common_names[[131]]$name
#> [1] "Себ"
#> 
#> $common_names[[131]]$language
#> [1] "tg"
#> 
#> 
#> $common_names[[132]]
#> $common_names[[132]]$name
#> [1] "แอปเปิล"
#> 
#> $common_names[[132]]$language
#> [1] "th"
#> 
#> 
#> $common_names[[133]]
#> $common_names[[133]]$name
#> [1] "Mansanas"
#> 
#> $common_names[[133]]$language
#> [1] "tl"
#> 
#> 
#> $common_names[[134]]
#> $common_names[[134]]$name
#> [1] "Elma"
#> 
#> $common_names[[134]]$language
#> [1] "tr"
#> 
#> 
#> $common_names[[135]]
#> $common_names[[135]]$name
#> [1] "Алма"
#> 
#> $common_names[[135]]$language
#> [1] "tt"
#> 
#> 
#> $common_names[[136]]
#> $common_names[[136]]$name
#> [1] "Улмопу"
#> 
#> $common_names[[136]]$language
#> [1] "udm"
#> 
#> 
#> $common_names[[137]]
#> $common_names[[137]]$name
#> [1] "ئالما"
#> 
#> $common_names[[137]]$language
#> [1] "ug"
#> 
#> 
#> $common_names[[138]]
#> $common_names[[138]]$name
#> [1] "Яблуня райська, Яблуко"
#> 
#> $common_names[[138]]$language
#> [1] "uk"
#> 
#> 
#> $common_names[[139]]
#> $common_names[[139]]$name
#> [1] "سیب"
#> 
#> $common_names[[139]]$language
#> [1] "ur"
#> 
#> 
#> $common_names[[140]]
#> $common_names[[140]]$name
#> [1] "Olma"
#> 
#> $common_names[[140]]$language
#> [1] "uz"
#> 
#> 
#> $common_names[[141]]
#> $common_names[[141]]$name
#> [1] "Apula"
#> 
#> $common_names[[141]]$language
#> [1] "ve"
#> 
#> 
#> $common_names[[142]]
#> $common_names[[142]]$name
#> [1] "Pomaro"
#> 
#> $common_names[[142]]$language
#> [1] "vec"
#> 
#> 
#> $common_names[[143]]
#> $common_names[[143]]$name
#> [1] "Táo tây"
#> 
#> $common_names[[143]]$language
#> [1] "vi"
#> 
#> 
#> $common_names[[144]]
#> $common_names[[144]]$name
#> [1] "Uibu"
#> 
#> $common_names[[144]]$language
#> [1] "vro"
#> 
#> 
#> $common_names[[145]]
#> $common_names[[145]]$name
#> [1] "Pemî"
#> 
#> $common_names[[145]]$language
#> [1] "wa"
#> 
#> 
#> $common_names[[146]]
#> $common_names[[146]]$name
#> [1] "Mansanas"
#> 
#> $common_names[[146]]$language
#> [1] "war"
#> 
#> 
#> $common_names[[147]]
#> $common_names[[147]]$name
#> [1] "עפל"
#> 
#> $common_names[[147]]$language
#> [1] "yi"
#> 
#> 
#> $common_names[[148]]
#> $common_names[[148]]$name
#> [1] "蘋果, 苹果"
#> 
#> $common_names[[148]]$language
#> [1] "zh"
#> 
#> 
#> $common_names[[149]]
#> $common_names[[149]]$name
#> [1] "Phōng-kó"
#> 
#> $common_names[[149]]$language
#> [1] "zh-min-nan"
#> 
#> 
#> $common_names[[150]]
#> $common_names[[150]]$name
#> [1] "蘋果"
#> 
#> $common_names[[150]]$language
#> [1] "zh-tw"
#> 
#> 
#> $common_names[[151]]
#> $common_names[[151]]$name
#> [1] "蘋果"
#> 
#> $common_names[[151]]$language
#> [1] "zh-yue"
#> 
#> 
#> 
#> $classification
#> $classification[[1]]
#> $classification[[1]]$rank
#> [1] "Domain"
#> 
#> $classification[[1]]$name
#> [1] "Eukaryota"
#> 
#> 
#> $classification[[2]]
#> $classification[[2]]$rank
#> [1] "• unranked"
#> 
#> $classification[[2]]$name
#> [1] "Archaeplastida"
#> 
#> 
#> $classification[[3]]
#> $classification[[3]]$rank
#> [1] "• Regnum"
#> 
#> $classification[[3]]$name
#> [1] "Plantae"
#> 
#> 
#> $classification[[4]]
#> $classification[[4]]$rank
#> [1] "• Cladus"
#> 
#> $classification[[4]]$name
#> [1] "angiosperms"
#> 
#> 
#> $classification[[5]]
#> $classification[[5]]$rank
#> [1] "• Cladus"
#> 
#> $classification[[5]]$name
#> [1] "eudicots"
#> 
#> 
#> $classification[[6]]
#> $classification[[6]]$rank
#> [1] "• Cladus"
#> 
#> $classification[[6]]$name
#> [1] "core eudicots"
#> 
#> 
#> $classification[[7]]
#> $classification[[7]]$rank
#> [1] "• Cladus"
#> 
#> $classification[[7]]$name
#> [1] "superrosids"
#> 
#> 
#> $classification[[8]]
#> $classification[[8]]$rank
#> [1] "• Cladus"
#> 
#> $classification[[8]]$name
#> [1] "rosids"
#> 
#> 
#> $classification[[9]]
#> $classification[[9]]$rank
#> [1] "• Cladus"
#> 
#> $classification[[9]]$name
#> [1] "eurosids I"
#> 
#> 
#> $classification[[10]]
#> $classification[[10]]$rank
#> [1] "• Ordo"
#> 
#> $classification[[10]]$name
#> [1] "Rosales"
#> 
#> 
#> $classification[[11]]
#> $classification[[11]]$rank
#> [1] "• Familia"
#> 
#> $classification[[11]]$name
#> [1] "Rosaceae"
#> 
#> 
#> $classification[[12]]
#> $classification[[12]]$rank
#> [1] "• Subfamilia"
#> 
#> $classification[[12]]$name
#> [1] "Spiraeoideae"
#> 
#> 
#> $classification[[13]]
#> $classification[[13]]$rank
#> [1] "• Tribus"
#> 
#> $classification[[13]]$name
#> [1] "Pyreae"
#> 
#> 
#> $classification[[14]]
#> $classification[[14]]$rank
#> [1] "• Genus"
#> 
#> $classification[[14]]$name
#> [1] "Malus"
#> 
#> 
#> $classification[[15]]
#> $classification[[15]]$rank
#> [1] "•"
#> 
#> $classification[[15]]$name
#> [1] "Malus domestica"
#> 
#> 
#> $classification[[16]]
#> $classification[[16]]$rank
#> [1] ""
#> 
#> $classification[[16]]$name
#> [1] "Borkh. (1803)"
#> 
#> 
#> 

# search wikicommons
wt_wikicommons_search(query = "Pinus")

#> $batchcomplete
#> [1] ""
#> 
#> $continue
#> $continue$sroffset
#> [1] 10
#> 
#> $continue$continue
#> [1] "-||"
#> 
#> 
#> $query
#> $query$searchinfo
#> $query$searchinfo$totalhits
#> [1] 257
#> 
#> 
#> $query$search
#> # A tibble: 10 × 6
#>       ns                                    title  size wordcount
#> *  <int>                                    <chr> <int>     <int>
#> 1      0                                    Pinus  4160       297
#> 2      0                              Pinus nigra  7449       486
#> 3      0                             Spinus pinus  1563       242
#> 4      0                       Pinus × schwerinii   634        65
#> 5      0                               Pinus mugo  7157       571
#> 6      0                       Pinus tabuliformis  1739       136
#> 7      0                          Setophaga pinus  1735       196
#> 8      0 Pinus distribution maps of North America 25971        92
#> 9      0                          Pinus sabiniana  2799       215
#> 10     0                            Pinus cooperi   564        62
#> # ... with 2 more variables: snippet <chr>, timestamp <chr>
#> 
#> 

## use search results to dig into pages
res <- wt_wikicommons_search(query = "Pinus")
lapply(res$query$search$title[1:3], wt_wikicommons)

#> [[1]]
#> [[1]]$langlinks
#> # A tibble: 103 × 5
#>       lang                                                          url
#> *    <chr>                                                        <chr>
#> 1       af                      https://af.wikipedia.org/wiki/Denneboom
#> 2      als                 https://als.wikipedia.org/wiki/Kiefer_(Baum)
#> 3       an                          https://an.wikipedia.org/wiki/Pinus
#> 4      ang            https://ang.wikipedia.org/wiki/P%C4%ABntr%C4%93ow
#> 5       ar https://ar.wikipedia.org/wiki/%D8%B5%D9%86%D9%88%D8%A8%D8%B1
#> 6       az       https://az.wikipedia.org/wiki/%C5%9Eam_a%C4%9Fac%C4%B1
#> 7      bar                        https://bar.wikipedia.org/wiki/Farchn
#> 8  bat-smg                https://bat-smg.wikipedia.org/wiki/Po%C5%A1is
#> 9       be       https://be.wikipedia.org/wiki/%D0%A5%D0%B2%D0%BE%D1%8F
#> 10      bg             https://bg.wikipedia.org/wiki/%D0%91%D0%BE%D1%80
#> # ... with 93 more rows, and 3 more variables: langname <chr>, autonym <chr>,
#> #   `*` <chr>
#> 
#> [[1]]$externallinks
#> [1] "https://www.wikidata.org/wiki/Q12024"
#> 
#> [[1]]$common_names
#> # A tibble: 101 × 2
#>         name language
#>        <chr>    <chr>
#> 1       Pine       en
#> 2  Denneboom       af
#> 3     Kiefer      als
#> 4   Pīntrēoƿ      ang
#> 5      صنوبر       ar
#> 6  Şam ağacı       az
#> 7     Kiefan      bar
#> 8       Хвоя       be
#> 9        Бор       bg
#> 10   ཐང་ཤིང་།       bo
#> # ... with 91 more rows
#> 
#> [[1]]$classification
#> # A tibble: 0 × 0
#> 
#> 
#> [[2]]
#> [[2]]$langlinks
#> # A tibble: 48 × 5
#>     lang
#> *  <chr>
#> 1     an
#> 2     ar
#> 3     az
#> 4     be
#> 5     bg
#> 6     ca
#> 7    ceb
#> 8     co
#> 9     cs
#> 10   csb
#> # ... with 38 more rows, and 4 more variables: url <chr>, langname <chr>,
#> #   autonym <chr>, `*` <chr>
#> 
#> [[2]]$externallinks
#> list()
#> 
#> [[2]]$common_names
#> # A tibble: 14 × 2
#>                             name language
#>                            <chr>    <chr>
#> 1            European Black Pine       en
#> 2                        Pinassa       ca
#> 3                  Schwarzkiefer       de
#> 4                 Pino salgareño       es
#> 5                    Feketefenyő       hu
#> 6             ヨーロッパクロマツ       ja
#> 7                        Црн бор       mk
#> 8                     Zwarte den       nl
#> 9                   Sosna czarna       pl
#> 10           Crni bor / Црни бор       sh
#> 11               Borovica čierna       sk
#> 12                       Karaçam       tr
#> 13                 Austrian Pine       en
#> 14 Corsican Pine, Pyreneean Pine       en
#> 
#> [[2]]$classification
#> # A tibble: 0 × 0
#> 
#> 
#> [[3]]
#> [[3]]$langlinks
#> # A tibble: 23 × 5
#>     lang                                                 url  langname
#> *  <chr>                                               <chr>     <chr>
#> 1    ast         https://ast.wikipedia.org/wiki/Spinus_pinus  Asturian
#> 2     bg          https://bg.wikipedia.org/wiki/Spinus_pinus Bulgarian
#> 3     ca       https://ca.wikipedia.org/wiki/Lluer_dels_pins   Catalan
#> 4    ceb      https://ceb.wikipedia.org/wiki/Carduelis_pinus   Cebuano
#> 5     cy https://cy.wikipedia.org/wiki/Pila_gwyrdd_y_pinwydd     Welsh
#> 6     de         https://de.wikipedia.org/wiki/Fichtenzeisig    German
#> 7     en           https://en.wikipedia.org/wiki/Pine_siskin   English
#> 8     eo            https://eo.wikipedia.org/wiki/Pinkardelo Esperanto
#> 9     es          https://es.wikipedia.org/wiki/Spinus_pinus   Spanish
#> 10    eu       https://eu.wikipedia.org/wiki/Carduelis_pinus    Basque
#> # ... with 13 more rows, and 2 more variables: autonym <chr>, `*` <chr>
#> 
#> [[3]]$externallinks
#> [1] "https://www.wikidata.org/wiki/Q26847063"                 
#> [2] "http://www.worldbirdnames.org"                           
#> [3] "http://www.worldbirdnames.org/bow/finches/"              
#> [4] "http://www.iucnredlist.org/apps/redlist/details/22720359"
#> 
#> [[3]]$common_names
#> # A tibble: 30 × 2
#>                                    name language
#>                                   <chr>    <chr>
#> 1                           Pine Siskin       en
#> 2                       lluer dels pins       ca
#> 3                       Carduelis pinus      ceb
#> 4                  čížek severoamerický       cs
#> 5                 Pila gwyrdd y pinwydd       cy
#> 6                           Fyrresisken       da
#> 7                         Fichtenzeisig       de
#> 8                            Pinkardelo       eo
#> 9  Jilguero de los Pinos, Pinero rayado       es
#> 10                         männisiisike       et
#> # ... with 20 more rows
#> 
#> [[3]]$classification
#> # A tibble: 0 × 0
#> 
#>