WikiSpecies

WikiSpecies

wt_wikispecies(name, utf8 = TRUE, ...)

wt_wikispecies_parse(page, types = c("langlinks", "iwlinks", "externallinks",
  "common_names", "classification"), tidy = FALSE)

wt_wikispecies_search(query, limit = 10, offset = 0, utf8 = TRUE, ...)

Arguments

name

(character) Wiki name - as a page title, must be length 1

utf8

(logical) If `TRUE`, encodes most (but not all) non-ASCII characters as UTF-8 instead of replacing them with hexadecimal escape sequences. Default: `TRUE`

...

curl options, passed on to [httr::GET()]

page

([httr::response()]) Result of [wt_wiki_page()]

types

(character) List of properties to parse

tidy

(logical). tidy output to data.frame's if possible. Default: `FALSE`

query

(character) query terms

limit

(integer) number of results to return. Default: 10

offset

(integer) record to start at. Default: 0

Value

wt_wikispecies returns a list, with slots:

langlinks - language page links

externallinks - external links

common_names - a data.frame with name and language columns

classification - a data.frame with rank and name columns

wt_wikispecies_parse returns a list

wt_wikispecies_search returns a list with slots for continue and query, where query holds the results, with query$search slot with the search results

References

https://www.mediawiki.org/wiki/API:Search for help on search

Examples

# high level
wt_wikispecies(name = "Malus domestica")

#> $langlinks
#> # A tibble: 1 × 5
#>    lang                                           url    langname     autonym
#> * <chr>                                         <chr>       <chr>       <chr>
#> 1    ia https://ia.wikipedia.org/wiki/Malus_domestica Interlingua interlingua
#> # ... with 1 more variables: `*` <chr>
#> 
#> $externallinks
#> [1] "http://www.ars-grin.gov/cgi-bin/npgs/html/taxon.pl?104681"
#> 
#> $common_names
#> # A tibble: 19 × 2
#>               name   language
#>              <chr>      <chr>
#> 1           Ябълка  български
#> 2     Poma, pomera     català
#> 3            Apfel    Deutsch
#> 4      Aed-õunapuu      eesti
#> 5            Μηλιά   Ελληνικά
#> 6            Apple    English
#> 7          Manzano    español
#> 8            Pomme   français
#> 9            Melâr     furlan
#> 10        사과나무     한국어
#> 11          ‘Āpala    Hawaiʻi
#> 12            Melo   italiano
#> 13           Aapel Nordfriisk
#> 14  Maçã, Macieira  português
#> 15 Яблоня домашняя    русский
#> 16   Tarhaomenapuu      suomi
#> 17            Elma     Türkçe
#> 18  Яблуня домашня українська
#> 19          Pomaro     vèneto
#> 
#> $classification
#> # A tibble: 8 × 2
#>          rank          name
#>         <chr>         <chr>
#> 1 Superregnum     Eukaryota
#> 2      Regnum       Plantae
#> 3      Cladus   Angiosperms
#> 4      Cladus      Eudicots
#> 5      Cladus Core eudicots
#> 6      Cladus        Rosids
#> 7      Cladus    Eurosids I
#> 8        Ordo       Rosales
#> 

# low level
pg <- wt_wiki_page("https://species.wikimedia.org/wiki/Abelmoschus")
wt_wikispecies_parse(pg)

#> $langlinks
#>  [1] "https://ar.wikipedia.org/wiki/%D9%85%D8%B3%D9%83%D9%8A"                                                                           
#>  [2] "https://az.wikipedia.org/wiki/Abelmo%C5%9F"                                                                                       
#>  [3] "https://ca.wikipedia.org/wiki/Abelmoschus"                                                                                        
#>  [4] "https://ceb.wikipedia.org/wiki/Abelmoschus"                                                                                       
#>  [5] "https://cs.wikipedia.org/wiki/Ibi%C5%A1kovec"                                                                                     
#>  [6] "https://de.wikipedia.org/wiki/Bisameibisch"                                                                                       
#>  [7] "https://en.wikipedia.org/wiki/Abelmoschus"                                                                                        
#>  [8] "https://eo.wikipedia.org/wiki/Abelmoschus"                                                                                        
#>  [9] "https://es.wikipedia.org/wiki/Abelmoschus"                                                                                        
#> [10] "https://fi.wikipedia.org/wiki/Okrat"                                                                                              
#> [11] "https://fr.wikipedia.org/wiki/Abelmoschus"                                                                                        
#> [12] "https://ga.wikipedia.org/wiki/Abelmoschus"                                                                                        
#> [13] "https://gl.wikipedia.org/wiki/Abelmosco"                                                                                          
#> [14] "https://hr.wikipedia.org/wiki/Abelmo%C5%A1us"                                                                                     
#> [15] "https://ia.wikipedia.org/wiki/Abelmoschus"                                                                                        
#> [16] "https://lt.wikipedia.org/wiki/Ybi%C5%A1k%C4%97"                                                                                   
#> [17] "https://ml.wikipedia.org/wiki/%E0%B4%85%E0%B4%AC%E0%B5%86%E0%B5%BD%E0%B4%AE%E0%B5%8B%E0%B4%B8%E0%B5%8D%E0%B4%95%E0%B4%B8%E0%B5%8D"
#> [18] "https://mrj.wikipedia.org/wiki/%D0%90%D0%B1%D0%B5%D0%BB%D1%8C%D0%BC%D0%BE%D1%88"                                                  
#> [19] "https://nl.wikipedia.org/wiki/Abelmoschus"                                                                                        
#> [20] "https://pl.wikipedia.org/wiki/Pi%C5%BCmian"                                                                                       
#> [21] "https://pt.wikipedia.org/wiki/Abelmoschus"                                                                                        
#> [22] "https://ru.wikipedia.org/wiki/%D0%90%D0%B1%D0%B5%D0%BB%D1%8C%D0%BC%D0%BE%D1%88"                                                   
#> [23] "https://sd.wikipedia.org/wiki/%D9%85%D9%88%D8%B1%D9%8A"                                                                           
#> [24] "https://sv.wikipedia.org/wiki/Okrasl%C3%A4ktet"                                                                                   
#> [25] "https://udm.wikipedia.org/wiki/%D0%90%D0%B1%D0%B5%D0%BB%D1%8C%D0%BC%D0%BE%D1%88"                                                  
#> [26] "https://vi.wikipedia.org/wiki/Chi_V%C3%B4ng_vang"                                                                                 
#> [27] "https://war.wikipedia.org/wiki/Abelmoschus"                                                                                       
#> [28] "https://zh.wikipedia.org/wiki/%E9%BB%84%E8%91%B5%E5%B1%9E"                                                                        
#> 
#> $iwlinks
#> [1] "https://en.wikipedia.org/wiki/Type_species"             
#> [2] "https://commons.wikimedia.org/wiki/Category:Abelmoschus"
#> [3] "https://commons.wikimedia.org/wiki/"                    
#> 
#> $externallinks
#> $externallinks[[1]]
#> [1] "http://www.malvaceae.info/Genera/Abelmoschus/Abelmoschus.php"
#> 
#> $externallinks[[2]]
#> [1] "http://mansfeld.ipk-gatersleben.de/pls/htmldb_pgrc/f?p=185:3:0::NO"
#> 
#> 
#> $common_names
#> $common_names[[1]]
#> $common_names[[1]]$name
#> [1] "Okrat"
#> 
#> $common_names[[1]]$language
#> [1] "suomi"
#> 
#> 
#> 
#> $classification
#> $classification[[1]]
#> $classification[[1]]$rank
#> [1] "Superregnum"
#> 
#> $classification[[1]]$name
#> [1] "Eukaryota"
#> 
#> 
#> $classification[[2]]
#> $classification[[2]]$rank
#> [1] "Regnum"
#> 
#> $classification[[2]]$name
#> [1] "Plantae"
#> 
#> 
#> $classification[[3]]
#> $classification[[3]]$rank
#> [1] "Cladus"
#> 
#> $classification[[3]]$name
#> [1] "Angiosperms"
#> 
#> 
#> $classification[[4]]
#> $classification[[4]]$rank
#> [1] "Cladus"
#> 
#> $classification[[4]]$name
#> [1] "Eudicots"
#> 
#> 
#> $classification[[5]]
#> $classification[[5]]$rank
#> [1] "Cladus"
#> 
#> $classification[[5]]$name
#> [1] "Core eudicots"
#> 
#> 
#> $classification[[6]]
#> $classification[[6]]$rank
#> [1] "Cladus"
#> 
#> $classification[[6]]$name
#> [1] "Rosids"
#> 
#> 
#> $classification[[7]]
#> $classification[[7]]$rank
#> [1] "Cladus"
#> 
#> $classification[[7]]$name
#> [1] "Eurosids II"
#> 
#> 
#> $classification[[8]]
#> $classification[[8]]$rank
#> [1] "Ordo"
#> 
#> $classification[[8]]$name
#> [1] "Malvales"
#> 
#> 
#> 

# search wikispecies
wt_wikispecies_search(query = "pine tree")

#> $batchcomplete
#> [1] ""
#> 
#> $continue
#> $continue$sroffset
#> [1] 10
#> 
#> $continue$continue
#> [1] "-||"
#> 
#> 
#> $query
#> $query$searchinfo
#> $query$searchinfo$totalhits
#> [1] 37
#> 
#> $query$searchinfo$suggestion
#> [1] "ping true"
#> 
#> $query$searchinfo$suggestionsnippet
#> [1] "<em>ping true</em>"
#> 
#> 
#> $query$search
#> # A tibble: 10 × 6
#>       ns                    title  size wordcount
#> *  <int>                    <chr> <int>     <int>
#> 1      0           Oecanthus pini   215        20
#> 2      0   Melanoplus punctulatus   689        63
#> 3      0         Dendrolimus pini   615        61
#> 4      0            Pineus laevis   940        86
#> 5      0 Brachytydeus danielssoni   733       164
#> 6      0  Casuarina equisetifolia   897       104
#> 7      0                   Pineus  2143       225
#> 8      0        Newnhamia petiola  2073       222
#> 9      0        Bennelongia pinpi  2191       243
#> 10     0       Nazeris spiculatus   595       107
#> # ... with 2 more variables: snippet <chr>, timestamp <chr>
#> 
#> 

## use search results to dig into pages
res <- wt_wikispecies_search(query = "pine tree")
lapply(res$query$search$title[1:3], wt_wikispecies)

#> [[1]]
#> [[1]]$langlinks
#> # A tibble: 4 × 5
#>    lang                                           url langname    autonym
#> * <chr>                                         <chr>    <chr>      <chr>
#> 1   ceb https://ceb.wikipedia.org/wiki/Oecanthus_pini  Cebuano    Cebuano
#> 2    nl  https://nl.wikipedia.org/wiki/Oecanthus_pini    Dutch Nederlands
#> 3    sv  https://sv.wikipedia.org/wiki/Oecanthus_pini  Swedish    svenska
#> 4   war https://war.wikipedia.org/wiki/Oecanthus_pini    Waray    Winaray
#> # ... with 1 more variables: `*` <chr>
#> 
#> [[1]]$externallinks
#> [1] "http://data.gbif.org/species/14194484"
#> 
#> [[1]]$common_names
#> # A tibble: 1 × 2
#>                name language
#>               <chr>    <chr>
#> 1 Pine Tree Cricket  English
#> 
#> [[1]]$classification
#> # A tibble: 16 × 2
#>            rank          name
#>           <chr>         <chr>
#> 1   Superregnum     Eukaryota
#> 2        Regnum      Animalia
#> 3     Subregnum     Eumetazoa
#> 4        Cladus     Bilateria
#> 5        Cladus     Nephrozoa
#> 6        Cladus   Protostomia
#> 7        Cladus     Ecdysozoa
#> 8        Cladus Panarthropoda
#> 9        Phylum    Arthropoda
#> 10    Subphylum      Hexapoda
#> 11      Classis       Insecta
#> 12       Cladus    Dicondylia
#> 13   Subclassis     Pterygota
#> 14         Ordo    Orthoptera
#> 15      Subordo      Ensifera
#> 16 Superfamilia    Grylloidea
#> 
#> 
#> [[2]]
#> [[2]]$langlinks
#> # A tibble: 4 × 5
#>    lang                                                   url langname
#> * <chr>                                                 <chr>    <chr>
#> 1   ceb https://ceb.wikipedia.org/wiki/Melanoplus_punctulatus  Cebuano
#> 2    nl  https://nl.wikipedia.org/wiki/Melanoplus_punctulatus    Dutch
#> 3    sv  https://sv.wikipedia.org/wiki/Melanoplus_punctulatus  Swedish
#> 4   war https://war.wikipedia.org/wiki/Melanoplus_punctulatus    Waray
#> # ... with 2 more variables: autonym <chr>, `*` <chr>
#> 
#> [[2]]$externallinks
#> list()
#> 
#> [[2]]$common_names
#> # A tibble: 1 × 2
#>                                name language
#>                               <chr>    <chr>
#> 1 Pine Tree Spur-throat Grasshopper  English
#> 
#> [[2]]$classification
#> # A tibble: 17 × 2
#>            rank          name
#>           <chr>         <chr>
#> 1   Superregnum     Eukaryota
#> 2        Regnum      Animalia
#> 3     Subregnum     Eumetazoa
#> 4        Cladus     Bilateria
#> 5        Cladus     Nephrozoa
#> 6        Cladus   Protostomia
#> 7        Cladus     Ecdysozoa
#> 8        Cladus Panarthropoda
#> 9        Phylum    Arthropoda
#> 10    Subphylum      Hexapoda
#> 11      Classis       Insecta
#> 12       Cladus    Dicondylia
#> 13   Subclassis     Pterygota
#> 14         Ordo    Orthoptera
#> 15      Subordo     Caelifera
#> 16    Infraordo     Acrididea
#> 17 Superfamilia    Acridoidea
#> 
#> 
#> [[3]]
#> [[3]]$langlinks
#> # A tibble: 19 × 5
#>     lang
#> *  <chr>
#> 1     be
#> 2     ca
#> 3     cs
#> 4     cy
#> 5     de
#> 6     en
#> 7     fi
#> 8     fr
#> 9     ka
#> 10    lv
#> 11    nl
#> 12    no
#> 13    pl
#> 14   pms
#> 15    pt
#> 16    ru
#> 17    sv
#> 18    uk
#> 19    vi
#> # ... with 4 more variables: url <chr>, langname <chr>, autonym <chr>,
#> #   `*` <chr>
#> 
#> [[3]]$externallinks
#> list()
#> 
#> [[3]]$common_names
#> # A tibble: 2 × 2
#>               name language
#>              <chr>    <chr>
#> 1 Pine-tree Lappet  English
#> 2    Bombyx du pin français
#> 
#> [[3]]$classification
#> # A tibble: 29 × 2
#>           rank          name
#>          <chr>         <chr>
#> 1  Superregnum     Eukaryota
#> 2       Regnum      Animalia
#> 3    Subregnum     Eumetazoa
#> 4       Cladus     Bilateria
#> 5       Cladus     Nephrozoa
#> 6       Cladus   Protostomia
#> 7       Cladus     Ecdysozoa
#> 8       Cladus Panarthropoda
#> 9       Phylum    Arthropoda
#> 10   Subphylum      Hexapoda
#> # ... with 19 more rows
#> 
#>